requirements.txt
opentelemetry-api==1.21.0
opentelemetry-distro==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0

